+++ eslint.config.js
@@ -5,9 +5,9 @@ module.exports = [
   {
     rules: {
       // Place any rule overrides in here
-    }
+    },
   },
   {
     ignores: ['build', 'dist', 'eslint.config.*', 'prettier.config.*', 'src/generated/*'],
-  }
-]
+  },
+];