+++ src/components/FloatingActionButton/_FloatingActionButton.context.ts
@@ -1,4 +1,4 @@
-import { createContext } from 'react';
+import { createContext, Dispatch, SetStateAction } from 'react';
 import { IconDefinition } from '@fortawesome/fontawesome-svg-core';
 import { ColourVariant } from '../../theme/theme.types';
 
@@ -7,17 +7,29 @@ export interface FabProps {
   icon: IconDefinition | null;
   onClick: () => void;
   variant?: ColourVariant;
+  label?: string;
+  dataTestId?: string;
+}
+
+export interface SecondaryFabProps {
+  id: string;
+  icon: IconDefinition;
+  onClick: () => void;
+  variant?: ColourVariant;
+  label?: string;
   dataTestId?: string;
 }
 
 interface FloatingActionButtonContextProps {
   contextExists: boolean;
-  setButton: (props: FabProps | undefined) => void;
+  setPrimaryButton: Dispatch<SetStateAction<FabProps | undefined>>;
+  setSecondaryButtons: Dispatch<SetStateAction<SecondaryFabProps[]>>;
 }
 
 const FloatingActionButtonContext = createContext<FloatingActionButtonContextProps>({
   contextExists: false,
-  setButton: () => {},
+  setPrimaryButton: () => {},
+  setSecondaryButtons: () => {},
 });
 
 export default FloatingActionButtonContext;