+++ src/components/FloatingActionButton/_FloatingActionButton.provider.tsx
@@ -1,36 +1,60 @@
-import React, { useMemo, useState } from 'react';
+import React, { useContext, useMemo, useState } from 'react';
 import { AnimatePresence } from 'framer-motion';
-import FloatingActionButtonContext, { FabProps } from './_FloatingActionButton.context';
-import FloatingActionButtonInternal from './_FloatingActionButton.internal';
+import FloatingActionButtonContext, { FabProps, SecondaryFabProps } from './_FloatingActionButton.context';
+import FloatingActionButtonInternal, { MiniFabInternal } from './_FloatingActionButton.internal';
+import MinimalMenuContext from '../MinimalMenu/MinimalMenu.context';
 
 export interface FloatingActionButtonProviderProps {
   children: React.ReactNode;
 }
 
 const FloatingActionButtonProvider = ({ children }: FloatingActionButtonProviderProps) => {
-  const [fab, setFab] = useState<FabProps | undefined>();
+  const [primaryFab, setPrimaryFab] = useState<FabProps | undefined>();
+  const [secondaryFabs, setSecondaryFabs] = useState<SecondaryFabProps[]>([]);
+  const { menuExists, isMobile } = useContext(MinimalMenuContext);
 
   const contextVal = useMemo(
     () => ({
       contextExists: true,
-      setButton: setFab,
+      setPrimaryButton: setPrimaryFab,
+      setSecondaryButtons: setSecondaryFabs,
     }),
-    [setFab],
+    [],
   );
 
   return (
     <FloatingActionButtonContext.Provider value={contextVal}>
       {children}
       <AnimatePresence>
-        {fab?.icon && (
+        {primaryFab?.icon && (
           <FloatingActionButtonInternal
-            key={fab.id}
-            icon={fab.icon}
-            onClick={fab.onClick}
-            variant={fab.variant}
-            data-testid={fab.dataTestId}
+            key={primaryFab.id}
+            icon={primaryFab.icon}
+            onClick={primaryFab.onClick}
+            variant={primaryFab.variant}
+            label={primaryFab.label}
+            data-testid={primaryFab.dataTestId}
           />
         )}
+        {secondaryFabs.map((fab, index) => {
+          const offsetBottom = menuExists && isMobile;
+          const baseOffset = offsetBottom ? 76 : 20;
+          const bottom = baseOffset + 56 + 8 + index * (40 + 10);
+          const staggerDelay = 0.2 + 0.05 * (index + 1);
+
+          return (
+            <MiniFabInternal
+              key={fab.id}
+              icon={fab.icon}
+              onClick={fab.onClick}
+              variant={fab.variant}
+              label={fab.label}
+              bottom={bottom}
+              staggerDelay={staggerDelay}
+              data-testid={fab.dataTestId}
+            />
+          );
+        })}
       </AnimatePresence>
     </FloatingActionButtonContext.Provider>
   );